Collection/LiveCollector/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0 